cleanup: move pathutil into its own package
