+++ engine/client/client.go
@@ -47,6 +47,7 @@ import (
 	"github.com/dagger/dagger/analytics"
 	"github.com/dagger/dagger/engine"
 	"github.com/dagger/dagger/engine/client/drivers"
+	"github.com/dagger/dagger/engine/client/pathutil"
 	"github.com/dagger/dagger/engine/client/secretprovider"
 	"github.com/dagger/dagger/engine/session"
 	"github.com/dagger/dagger/engine/slog"
@@ -1090,7 +1091,7 @@ func (c *Client) clientMetadata() engine.ClientMetadata {
 	// for consistent behavior of CLI inside nested execution
 	homeDir, err := os.UserHomeDir()
 	if err == nil {
-		expandedPath, err := ExpandHomeDir(homeDir, sshAuthSock)
+		expandedPath, err := pathutil.ExpandHomeDir(homeDir, sshAuthSock)
 		if err == nil {
 			sshAuthSock = expandedPath
 		}
