+++ engine/client/pathutil/pathutil_test.go
@@ -1,4 +1,4 @@
-package client
+package pathutil
 
 import (
 	"testing"
